+++ src/main/java/org/sbrubbles/genericcons/C.java
@@ -1,9 +1,9 @@
 package org.sbrubbles.genericcons;
 
 /**
- * Captures and represents an open-ended list of types. This class isn't supposed to be instantiated or subclassed, 
- * and has no fields or methods; it's a marker class, used to indicate a list in the type declaration, like below:
- * <p>
+ * A marker class, used to build and represent an open-ended list of types in a generic declaration.
+ * Like below:
+ *
  * <pre>
  * Function&lt;?, ?&gt; f = new Function&lt;String, C&lt;Object, C&lt;Number, C&lt;String, Integer&gt;&gt;&gt;&gt;() {
  *   { // the no-arg constructor
@@ -12,13 +12,14 @@ package org.sbrubbles.genericcons;
  * };
  * </pre>
  * <p>
- * In this example, {@code C&lt;Object, C&lt;Number, C&lt;String, Integer&gt;&gt;&gt;&gt;} encodes that {@code Function}'s
- * second type argument is a list of types: {@code Object}, {@code Number}, {@code String} and {@code Integer}. The 
- * {@link Types} class provides static methods to extract the types. 
+ * {@code C<Object, C<Number, C<String, Integer>>>>} encodes a list of types:
+ * {@code Object}, {@code Number}, {@code String} and {@code Integer}. The {@link Types} class provides methods to
+ * extract the types for later use.
  * <p>
- * This list-like structure, similar to Lisp's {@code cons}, is named a <i>cons type</i> (hence the {@code C}), 
- * but using only one letter kept the whole structure more readable.
+ * This list-like structure is named a <i>cons type</i> (hence the {@code C}), due to its similarities to
+ * Lisp's {@code cons}. Using only one letter for the class kept the whole structure more readable.
  * <p>
+ * This class isn't supposed to be instantiated or subclassed, and has no fields or methods.
  * 
  * @author Humberto Anjos
  * @param <First> The first type.